refactor(clean-arch): split test evidence helpers so tests/core has no transitive framework imports
