+++ tests/core/test_no_framework_transitivity.py
@@ -2,14 +2,16 @@
 
 The clean-architecture self-evaluation flagged tests/core files for reaching
 httpx (via tests._evidence_helpers -> analysis.mcp.findings_server ->
-llm_bridge._embeddings) and pydantic (via data.events). Those chains were cut
-by deferring the analysis imports in tests/_evidence_helpers and by moving
-the event codec out of the core entities; this test keeps them cut.
+llm_bridge._embeddings) and pydantic (via analysis.mcp.handlers -> quodeq ->
+quodeq.cli). Those chains were cut by splitting the analysis-layer helpers
+out of tests/_evidence_helpers into tests/_analysis_helpers; this test keeps
+them cut, using both a runtime import probe and a static checker-based guard.
 """
 from __future__ import annotations
 
 import subprocess
 import sys
+from pathlib import Path
 
 _PROBE = """
 import sys
@@ -29,3 +31,16 @@ def test_core_test_helpers_import_no_frameworks():
         capture_output=True, text=True, timeout=120,
     )
     assert result.returncode == 0, result.stderr
+
+
+def test_tests_core_has_no_transitive_framework_dependency():
+    from quodeq.core.checks.framework_deps import check_framework_dependencies
+    from quodeq.core.checks.frameworks import FRAMEWORK_PACKAGES
+    from quodeq.data.fs.import_graph import build_import_graph
+
+    root = Path(__file__).resolve().parents[2]
+    files = [*(root / "src").rglob("*.py"), *(root / "tests").rglob("*.py")]
+    graph = build_import_graph(root, files)
+    judgments = check_framework_dependencies(graph, framework_packages=FRAMEWORK_PACKAGES, dimension="architecture")
+    offenders = [j for j in judgments if j.practice_id == "CLEA-DEP-06" and j.file.startswith("tests/core/")]
+    assert not offenders, [(j.file, j.reason) for j in offenders]
